cleanup EE folder and move everything to XEH
