--- addons/ee/$PBOPREFIX$
@@ -1,1 +1,0 @@
-x\cba\addons\ee
