+++ addons/xeh/ee/config.cpp
@@ -1,5 +1,3 @@
-#include "script_component.hpp"
-
 class CfgPatches {
     class cba_ee {
         requiredAddons[] = {"cba_xeh"};
